+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/blocks
 
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.22
+
 ## 0.1.19
 
 ### Patch Changes