+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-ui-sdk
 
+## 1.4.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.22
+
 ## 1.4.18
 
 ### Patch Changes